==> sweden/1_merge_robust.ipynb <==
## Import

import pandas as pd
import numpy as np
import requests



import os
import sys
currentdir = os.path.dirname(os.path.realpath('analysis_DK'))
parentdir = os.path.dirname(currentdir)
sys.path.append(parentdir)

from func import global_id, quarter_to_month, reindex, GT_dict

## DK 

#### Target data 

df_dk_target = pd.read_csv(r'data/target/df_DK_target.csv', parse_dates = ['date'])

df_dk_target = df_dk_target.sort_values(by=['ID', 'date'])

df_dk_target.date.min()

df_dk_target.date.max()

#### Job posts

df_dk_jobs = pd.read_csv(r'data/job_posts/df_DK_jobposts_quarterly_final.csv', parse_dates = ['date'])

df_dk_jobs[df_dk_jobs.date == '2019-01-01'].jobs.sum()

df_dk_jobs = df_dk_jobs.drop(['labour_force'], axis = 1)

Merge on labour force 

df_dk_labour = pd.read_csv(r'data/job_posts/df_DK_labour_force.csv', parse_dates = ['date'])

df_dk_jobs = df_dk_jobs.merge(df_dk_labour, on = ['date', 'ID'])

df_dk_jobs['labour_force'] = df_dk_jobs.groupby('ID').labour_force.shift(1)

Create job rate

jobs_list = ['jobs', 'sector_information_technology', 'sector_engineering_technology','sector_management_staff', 'sector_trade_service',
       'sector_industry_craft', 'sector_sales_communication',
       'sector_teaching', 'sector_office_finance', 'sector_social_health',
       'sector_other']

for col in jobs_list:
    df_dk_jobs[col] = (df_dk_jobs[col] / df_dk_jobs['labour_force'])*100

df_dk_jobs = df_dk_jobs.drop(['labour_force'], axis = 1)

df_dk_jobs.date.min()

df_dk_jobs.date.max()

#### Google

#### Regional

df_dk_controls = pd.read_csv(r'data/controls/df_DK_controls.csv', parse_dates = ['date'])

df_dk_controls = df_dk_controls.drop(['w_ave_socio_index'], axis = 1)

df_dk_controls = df_dk_controls[df_dk_controls['date'].map(lambda x: x.month).isin([1, 4, 7, 10])]

df_dk_controls.date.min()

df_dk_controls.date.max()

#### Merge data

df_DK = df_dk_target.merge(df_dk_jobs, on = ['date', 'ID'])
df_DK.shape

df_DK = df_DK.merge(df_dk_controls, on = ['date', 'ID'])
df_DK.shape

df_DK.date.min()

df_DK[df_DK['jobs'].notnull()]

df_DK['country'] = 'DK'

df_DK.head()

df_DK.date.max()

df_DK.date.min()

#### Save the data 

df_DK.to_csv('data/DK_SE_NO/df_dk.csv')

## SE

#### Target data 

df_se_target = pd.read_csv(r'data/target/df_SE_target.csv', parse_dates = ['date'])

df_se_target.date.min()

df_se_target.date.max()

Create target lag

#df_se_target['target_lag'] = df_se_target.target_actual.shift(1)

#### Job posts

df_se_jobs = pd.read_csv(r'data/job_posts/df_SE_jobposts_quarterly_final.csv', parse_dates = ['date'])

df_se_jobs = df_se_jobs.drop(['labour_force'], axis = 1)

Merge on labour force 

df_se_labour = pd.read_csv(r'data/job_posts/df_SE_labour_force.csv', parse_dates = ['date'])

df_se_jobs = df_se_jobs.merge(df_se_labour, on = ['date', 'ID'])

df_se_jobs['labour_force'] = df_se_jobs.groupby('ID').labour_force.shift(1)

Create job rate

jobs_list = ['jobs', 'sector_information_technology', 'sector_engineering_technology','sector_management_staff', 'sector_trade_service',
       'sector_industry_craft', 'sector_sales_communication',
       'sector_teaching', 'sector_office_finance', 'sector_social_health',
       'sector_other']

for col in jobs_list:
    df_se_jobs[col] = (df_se_jobs[col] / df_se_jobs['labour_force'])*100

df_se_jobs = df_se_jobs.drop(['labour_force'], axis = 1)

df_se_jobs.date.min()

df_se_jobs.date.max()

#### Google

#### Regional

df_SE_controls = pd.read_csv('data/controls/df_SE_controls.csv', parse_dates = ['date'])

df_SE_controls = df_SE_controls[['reg_nr','ID','date','pop', 'labour_force_share','mvu_lvu_share','urban_share']]

df_SE_controls = df_SE_controls.replace('Västra Götaland', 'Västra Götalands')

df_SE_controls.date.min()

df_SE_controls.date.max()

Shift controls

df_SE_controls['pop'] = df_SE_controls.groupby('ID').pop.shift(5)
df_SE_controls['labour_force_share'] = df_SE_controls.groupby('ID').labour_force_share.shift(1)
df_SE_controls['mvu_lvu_share'] = df_SE_controls.groupby('ID').mvu_lvu_share.shift(6)
df_SE_controls['urban_share'] = df_SE_controls.groupby('ID').urban_share.shift(7)

#### Merge data

df_SE = df_se_target.merge(df_se_jobs, on = ['date', 'ID'])

df_SE = df_SE.merge(df_SE_controls, on = ['date', 'ID'], how = 'left')

df_SE['country'] = 'SE'

#df_SE = df_SE[df_SE.date >= '2007-01-01']

df_SE.date.min()

df_SE.date.max()

df_SE.shape

#### Save the data 

## NO

df_no_target = pd.read_csv(r'data/target/df_NO_target.csv',  parse_dates = ['date'])

#### Target data 

df_no_target.date.min()

df_no_target.date.max()

Create target lag

df_no_target['target_lag'] = df_no_target.groupby('ID').target_actual.shift(1)

#### Job posts

df_no_jobs = pd.read_csv(r'data/job_posts/df_NO_jobposts_quarterly_final.csv', parse_dates = ['date'])

df_no_jobs = df_no_jobs.drop(['labour_force'], axis = 1)

Correct Nord- Sør trøndelag

test = df_no_jobs[df_no_jobs['ID'].str.startswith('Nord-Trøndelag')]
#test = test.reset_index()

test2 = df_no_jobs[df_no_jobs['ID'].str.startswith('Sør-Trøndelag')]
#test2 = test2.reset_index()


temp = pd.concat([test,test2])
temp = temp.groupby('date').sum()
temp['ID'] ='Trøndelag'

temp = temp.reset_index()

Correct labour force for those two regions

df = pd.read_csv("data/other/df_NO_labour_force.csv",  parse_dates = ['date'])
df = df[['date','labour_force', 'ID']]

test = df[df['ID'].str.startswith('Nord-Trøndelag')]
test = test.reset_index()

test2 = df[df['ID'].str.startswith('Sør-Trøndelag')]
test2 = test2.reset_index()

test3 = df[df['ID'].str.startswith('Trøndelag')]
test3 = test3.reset_index()

temp = pd.concat([test,test2, test3])
temp = temp.groupby('date').sum()
temp['ID'] ='Trøndelag'

temp = temp.reset_index()

temp = temp[['date', 'labour_force', 'ID']]

df= df[~df.ID.isin(['Nord-Trøndelag', 'Sør-Trøndelag', 'Trøndelag'])]

df = pd.concat([df, temp])
df['labour_force'] = df['labour_force']*1000

df_no_jobs = df_no_jobs.merge(df, on = ['date', 'ID'])

df_no_jobs['labour_force'] =  df_no_jobs.groupby('ID').labour_force.shift(1)

Create job rate

jobs_list = ['jobs', 'sector_information_technology', 'sector_engineering_technology','sector_management_staff', 'sector_trade_service',
       'sector_industry_craft', 'sector_sales_communication',
       'sector_teaching', 'sector_office_finance', 'sector_social_health',
       'sector_other']

for col in jobs_list:
    df_no_jobs[col] = (df_no_jobs[col] / df_no_jobs['labour_force'])*100

df_no_jobs = df_no_jobs.drop(['labour_force'], axis = 1)

#### Google

#### Regional

df_NO_controls = pd.read_csv('data/other/df_NO_controls.csv', parse_dates = ['date'])

df_NO_controls.tail()

df_NO_controls = df_NO_controls.replace('Finnmark - Finnmárku', 'Finnmark')
df_NO_controls = df_NO_controls.replace('Troms - Romsa', 'Troms')

Shift controls

df_NO_controls['pop'] = df_NO_controls.groupby('ID').pop.shift(1)
df_NO_controls['labour_force_share'] = df_NO_controls.groupby('ID').labour_force_share.shift(1)
df_NO_controls['mvu_lvu_share'] = df_NO_controls.groupby('ID').mvu_lvu_share.shift(6)
df_NO_controls['urban_share'] = df_NO_controls.groupby('ID').urban_share.shift(4)

df_NO_controls.tail()

#### Merge data

df_NO = df_no_target.merge(df_no_jobs, on = ['date', 'ID'])

df_NO = df_NO.merge(df_NO_controls, on = ['date', 'ID'], how = 'left')

df_NO['country'] = 'NO'

df_NO = df_NO[df_NO.date >= '2007-01-01']

df_NO.date.min()

df_NO.date.max()

#### Save the data 

## Concat the files

df_DK = df_DK.rename(columns = {'pop' : 'population', 'mvu_lvu_share_pop' : 'high_edu_share',
       'w_ave_urban_index' : 'urban_share'})

df_DK = df_DK[['date', 'target_actual', 'ID','jobs',
       'sector_information_technology', 'sector_engineering_technology',
       'sector_management_staff', 'sector_trade_service',
       'sector_industry_craft', 'sector_sales_communication',
       'sector_teaching', 'sector_office_finance', 'sector_social_health',
       'sector_other', 'population', 'high_edu_share', 'labour_force_share',
       'urban_share', 'country']]

df_DK.columns

df_SE = df_SE.rename(columns = {'pop' : 'population', 'mvu_lvu_share' : 'high_edu_share'})

df_SE = df_SE[['date', 'target_actual', 'ID', 'jobs',
       'sector_information_technology', 'sector_engineering_technology',
       'sector_management_staff', 'sector_trade_service',
       'sector_industry_craft', 'sector_sales_communication',
       'sector_teaching', 'sector_office_finance', 'sector_social_health',
       'sector_other', 'population', 'high_edu_share', 'labour_force_share',
       'urban_share', 'country']]

df_SE.columns

df_NO = df_NO.rename(columns = {'pop' : 'population', 'mvu_lvu_share' : 'high_edu_share',
       'w_ave_urban_index' : 'urban_share'})

df_NO = df_NO[['date', 'target_actual', 'ID', 'target_lag', 'jobs',
       'sector_information_technology', 'sector_engineering_technology',
       'sector_management_staff', 'sector_trade_service',
       'sector_industry_craft', 'sector_sales_communication',
       'sector_teaching', 'sector_office_finance', 'sector_social_health',
       'sector_other', 'population', 'high_edu_share', 'labour_force_share',
       'urban_share', 'country']]

df_NO.columns

df_descriptive = pd.concat([df_DK, df_SE, df_NO])

df_descriptive

Secure that all are the same period

df_descriptive = df_descriptive[df_descriptive.date >= df_DK.date.min()]
df_descriptive = df_descriptive[df_descriptive.date <= df_DK.date.max()]

Check the data

df_descriptive.ID.value_counts()

df_descriptive.country.value_counts()

df_descriptive.isnull().sum()

temp = df_descriptive[df_descriptive.date.dt.year == 2010]
temp[temp.urban_share.isnull()]

df_descriptive.date.max()

df_descriptive.date.min()

Mean table

df_descriptive.groupby('country').mean()

## GT

dfTrends = pd.read_csv('data/gt/dfTrends.csv', index_col=0, sep = ';', parse_dates=['date'])

dfTrends.drop(['geo', 'trends', 'C_ID'], axis=1, inplace=True)

df_descriptive = df_DK.copy()

df_descriptive = df_descriptive[df_descriptive.date >= '01-01-2007']

df_descriptive.shape

df_descriptive = pd.merge(df_descriptive, dfTrends, how = 'left', left_on=['ID', 'date'], right_on=['ID', 'date'])

df_descriptive.shape

### Reindexing GT

#Reindexing trends
df_descriptive = df_descriptive.apply(lambda col: reindex(col) if col.name in GT_dict() else col).copy()

df_descriptive.columns

df_descriptive = df_descriptive.drop(['GT_NO_0', 'GT_NO_1', 'GT_NO_2', 'GT_NO_6', 'GT_NO_7', 'GT_NO_8', 'GT_NO_10', 'GT_NO_11', 
                                'GT_SE_0', 'GT_SE_1', 'GT_SE_2', 'GT_SE_5', 'GT_SE_6', 'GT_SE_7', 'GT_SE_8', 'GT_SE_9', 
                                'GT_SE_10', 'GT_SE_11'], axis = 1)

## Exporting

df_descriptive.to_csv('data/descriptive/df_DK_descriptive.csv', index = False)
