==> sweden/3_analysis_robust.ipynb <==
# Import 

# STANDARD PACKAGES
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
%matplotlib inline
import seaborn as sns
import re
from datetime import datetime as dt
import time
import json
import random
from tqdm import tqdm #https://pypi.org/project/tqdm/#ipython-jupyter-integ½ration
from functools import reduce
import pickle
import itertools
import warnings
import platform
import multiprocessing as mp


# SCRAPE PACKAGES
import requests
from bs4 import BeautifulSoup
# from pytrends.request import TrendReq #pip install pytrends

# MODEL PACKAGES
    #SKLEARN
from sklearn.linear_model import Ridge
from sklearn.linear_model import Lasso
from sklearn.linear_model import ElasticNet
from sklearn.ensemble import RandomForestRegressor
from sklearn.preprocessing import StandardScaler
from sklearn.preprocessing import PolynomialFeatures
from sklearn.decomposition import PCA
from statsmodels.tsa.x13 import x13_arima_analysis as x13

# CUSTOM FUNCTIONS

import os
import sys
currentdir = os.path.dirname(os.path.realpath('analysis_DK'))
parentdir = os.path.dirname(currentdir)
sys.path.append(parentdir)

from func import (chunks, reindex, global_id, term_list, time_corr_plot, rmse, time_variable_plot, find_highest_corr, test_train_split, test_train_split_Q,
                  bootstrap_all_windows, bootstrap_n_samples, bootstrap_sample, final_model, final_model_boot, ar_1, grid_bestpar, tuning_window, tuning_window_mp, tuning_window_bestpar,
                  model_tuning, seasadj, seasadj_col_list, abs_percentage_change, add_poly_terms, create_interaction, GT_dict)

# Final model

## Baseline -  AR(1) 

### Subset data 

df_analysis = pd.read_csv('data/descriptive/df_descriptive.csv', parse_dates=['date'])

df_analysis = df_analysis[df_analysis.country == 'SE']

df_analysis = df_analysis[~(df_analysis.ID == 'Jämtland')]

df_analysis = df_analysis.drop(['GT_NO_0', 'GT_NO_1', 'GT_NO_2', 'GT_NO_6', 'GT_NO_7', 'GT_NO_8', 'GT_NO_10', 'GT_NO_11',  
                                'GT_DK_0', 'GT_DK_2', 'GT_DK_3', 'GT_DK_4', 'GT_DK_5', 'GT_DK_6', 'GT_DK_7', 'GT_DK_8',
                                'GT_DK_9', 'GT_DK_10', 'GT_DK_11', 'GT_DK_12', 'GT_DK_13', 'GT_DK_14',
                                'GT_DK_15', 'GT_DK_16', 'GT_DK_17', 'GT_DK_18', 'GT_DK_19', 'GT_DK_20', 'GT_DK_21',
                                'country'], axis = 1)

df_analysis.reset_index(inplace = True)

df_analysis.to_csv('data/descriptive/df_analysis.csv', index = False)

df_analysis.groupby(['ID'])['population'].mean().round()

### Preprocessing

#### Transform relevant columns to abs change exept those with M_ and ID_, date and t

df_analysis['target_actual'] = df_analysis.groupby(['ID'])['target_actual'].diff()

### Lagged variables

df_analysis['target_lag'] = df_analysis.groupby(['ID'])['target_actual'].shift(1)

#### Creating dummies from categorial variables - remember to drop the reference category (done after change is constructed)

df_analysis = pd.get_dummies(df_analysis, prefix=['ID'], prefix_sep='_', columns=['ID']).copy()

#### Adding interaction terms by regions and all variables

# relevant interaction variables
interaction_1 = ['target_lag'] 

interaction_2 = [item for item in df_analysis if item.startswith('ID_')]

for var1 in interaction_1:
    for var2 in interaction_2:
        name = var1 + "*" + var2
        df_analysis[name] = pd.Series(df_analysis[var1] * df_analysis[var2], name=name)

#### Drop na

df_analysis.dropna(inplace=True)

df_analysis.sort_index(axis=1, inplace=True)

window = 11
testsize = 1
valsize = 1
rolling_window = True
params = []
n_components = []

#### Select relevant columns

df_ar = df_analysis[['date', 'target_actual', 
                     'target_lag*ID_Blekinge', 'target_lag*ID_Dalarna', 'target_lag*ID_Gotland', 'target_lag*ID_Gävleborg', 
                     'target_lag*ID_Halland', 'target_lag*ID_Jönköping', 'target_lag*ID_Kalmar', 
                     'target_lag*ID_Kronoberg', 'target_lag*ID_Norrbotten', 'target_lag*ID_Skåne', 'target_lag*ID_Stockholm', 
                     'target_lag*ID_Södermanland', 'target_lag*ID_Uppsala', 'target_lag*ID_Värmland', 'target_lag*ID_Västerbotten', 
                     'target_lag*ID_Västernorrland', 'target_lag*ID_Västmanland', 'target_lag*ID_Västra Götalands', 'target_lag*ID_Örebro', 
                     'target_lag*ID_Östergötland']]

#### To Quarter data

df_ar['date'] = df_ar.set_index('date').index.to_period('Q')

df_ar['date'].min()

df_ar['date'].max()

#### Subset for period to match DK

df_ar = df_ar[df_ar.date >= '01-01-2008']

### Running baseline model

X_train, X_val, X_test, y_train, y_val, y_test, y_dates = test_train_split_Q(df = df_ar, window = window, testsize=testsize, valsize = valsize,
                                                                  y_col='target_actual', rolling_window = rolling_window, df_output= False, geo_count = 20)

#### Concatting val and train

for win in X_train.keys():
    X_train[win] = np.concatenate((X_train[win], X_val[win])).copy()
    y_train[win] = np.concatenate((y_train[win], y_val[win])).copy()

#### Estimating with OLS

results_ols= tuning_window(X_fit = X_train, y_fit = y_train, X_test = X_test, y_test = y_test, params = params, n_components = n_components, model_str = 'ols')

results_ols[1]['best'][1]

#### Exporting the results

with open('results/final/baseline/results_ar1.pickle', 'wb') as handle:
    pickle.dump(results_ols, handle, protocol= pickle.HIGHEST_PROTOCOL)

### Check results

temp = []
for key in results_ols.keys():
    temp.append(results_ols[key]['best'][1])
    

np.mean(temp)

results_ols[1]['y_pred_dict'][results_ols[1]['best'][0]]

## Baseline -  AR(2) with y_t-1 + y_t-12

### Subset data 

df_analysis = pd.read_csv('data/descriptive/df_descriptive.csv', parse_dates=['date'])

df_analysis = df_analysis[df_analysis.country == 'SE']

df_analysis = df_analysis[~(df_analysis.ID == 'Jämtland')]

df_analysis = df_analysis.drop(['GT_NO_0', 'GT_NO_1', 'GT_NO_2', 'GT_NO_6', 'GT_NO_7', 'GT_NO_8', 'GT_NO_10', 'GT_NO_11',  
                                'GT_DK_0', 'GT_DK_2', 'GT_DK_3', 'GT_DK_4', 'GT_DK_5', 'GT_DK_6', 'GT_DK_7', 'GT_DK_8',
                                'GT_DK_9', 'GT_DK_10', 'GT_DK_11', 'GT_DK_12', 'GT_DK_13', 'GT_DK_14',
                                'GT_DK_15', 'GT_DK_16', 'GT_DK_17', 'GT_DK_18', 'GT_DK_19', 'GT_DK_20', 'GT_DK_21',
                                'country'], axis = 1)

### Preprocessing

#### Transform relevant columns to abs change exept those with M_ and ID_, date and t

df_analysis['target_actual'] = df_analysis.groupby(['ID'])['target_actual'].diff()

### Lagged variables

df_analysis['target_lag'] = df_analysis.groupby(['ID'])['target_actual'].shift(1)

df_analysis['target_4_lag'] = df_analysis.groupby(['ID'])['target_actual'].shift(4)

#### Creating dummies from categorial variables - remember to drop the reference category (done after change is constructed)

df_analysis = pd.get_dummies(df_analysis, prefix=['ID'], prefix_sep='_', columns=['ID']).copy()

#### Adding interaction terms by regions and all variables

# relevant interaction variables
interaction_1 = ['target_lag', 'target_4_lag'] 

interaction_2 = [item for item in df_analysis if item.startswith('ID_')]

for var1 in interaction_1:
    for var2 in interaction_2:
        name = var1 + "*" + var2
        df_analysis[name] = pd.Series(df_analysis[var1] * df_analysis[var2], name=name)

#### Drop na

df_analysis.dropna(inplace=True)

df_analysis.sort_index(axis=1, inplace=True)

window = 11
testsize = 1
valsize = 1
rolling_window = True
params = []
n_components = []

#### Select relevant columns

df_ar = df_analysis[['date', 'target_actual', 
                     'target_lag*ID_Blekinge', 'target_lag*ID_Dalarna', 'target_lag*ID_Gotland', 'target_lag*ID_Gävleborg', 
                     'target_lag*ID_Halland', 'target_lag*ID_Jönköping', 'target_lag*ID_Kalmar', 
                     'target_lag*ID_Kronoberg', 'target_lag*ID_Norrbotten', 'target_lag*ID_Skåne', 'target_lag*ID_Stockholm', 
                     'target_lag*ID_Södermanland', 'target_lag*ID_Uppsala', 'target_lag*ID_Värmland', 'target_lag*ID_Västerbotten', 
                     'target_lag*ID_Västernorrland', 'target_lag*ID_Västmanland', 'target_lag*ID_Västra Götalands', 'target_lag*ID_Örebro', 
                     'target_lag*ID_Östergötland',
                    'target_4_lag*ID_Blekinge', 'target_4_lag*ID_Dalarna', 'target_4_lag*ID_Gotland',
                     'target_4_lag*ID_Gävleborg', 'target_4_lag*ID_Halland', 'target_4_lag*ID_Jönköping', 
                     'target_4_lag*ID_Kalmar', 'target_4_lag*ID_Kronoberg', 'target_4_lag*ID_Norrbotten', 'target_4_lag*ID_Skåne', 
                     'target_4_lag*ID_Stockholm', 'target_4_lag*ID_Södermanland', 'target_4_lag*ID_Uppsala', 'target_4_lag*ID_Värmland',
                     'target_4_lag*ID_Västerbotten', 'target_4_lag*ID_Västernorrland', 'target_4_lag*ID_Västmanland', 
                     'target_4_lag*ID_Västra Götalands', 'target_4_lag*ID_Örebro', 'target_4_lag*ID_Östergötland']]

#### To Quarter data

df_ar['date'] = df_ar.set_index('date').index.to_period('Q')

df_ar['date'].min()

df_ar['date'].max()

#### Subset for period to match DK

df_ar = df_ar[df_ar.date >= '01-01-2008']

### Running baseline model

X_train, X_val, X_test, y_train, y_val, y_test, y_dates = test_train_split_Q(df = df_ar, window = window, testsize=testsize, valsize = valsize,
                                                                  y_col='target_actual', rolling_window = rolling_window, df_output= False, geo_count = 20)

#### Concatting val and train

for win in X_train.keys():
    X_train[win] = np.concatenate((X_train[win], X_val[win])).copy()
    y_train[win] = np.concatenate((y_train[win], y_val[win])).copy()

#### Estimating with OLS

results_ols= tuning_window(X_fit = X_train, y_fit = y_train, X_test = X_test, y_test = y_test, params = params, n_components = n_components, model_str = 'ols')

results_ols[1]['best'][1]

#### Exporting the results

with open('results/final/baseline/results_ar_year_lag.pickle', 'wb') as handle:
    pickle.dump(results_ols, handle, protocol= pickle.HIGHEST_PROTOCOL)

### Check results

temp = []
for key in results_ols.keys():
    temp.append(results_ols[key]['best'][1])
    

np.mean(temp)

results_ols[1]['y_pred_dict'][results_ols[1]['best'][0]]

## ML - Data and preprocessing 

### Import data frame with adjusted here

### Subset data 

df_analysis = pd.read_csv('data/descriptive/df_descriptive.csv', parse_dates=['date'])

df_analysis = df_analysis[df_analysis.country == 'SE']

df_analysis = df_analysis[~(df_analysis.ID == 'Jämtland')]

df_analysis = df_analysis.drop(['GT_NO_0', 'GT_NO_1', 'GT_NO_2', 'GT_NO_6', 'GT_NO_7', 'GT_NO_8', 'GT_NO_10', 'GT_NO_11',  
                                'GT_DK_0', 'GT_DK_2', 'GT_DK_3', 'GT_DK_4', 'GT_DK_5', 'GT_DK_6', 'GT_DK_7', 'GT_DK_8',
                                'GT_DK_9', 'GT_DK_10', 'GT_DK_11', 'GT_DK_12', 'GT_DK_13', 'GT_DK_14',
                                'GT_DK_15', 'GT_DK_16', 'GT_DK_17', 'GT_DK_18', 'GT_DK_19', 'GT_DK_20', 'GT_DK_21',
                                'country'], axis = 1)

print(list(df_analysis))

df_analysis.date.min()

df_analysis.date.max()

### Initial preprocessing and feature construction

- Create dummies 
- Create interaction terms

Overall monthly time trend variable, $t=1,2...,T$ within `ID` variable:

#Temp container
temp = {}

for i in df_analysis['ID'].unique():
    temp[i] = df_analysis[df_analysis['ID']==i]
    temp[i]['t'] = range(1, len(temp[i]['ID'])+1)

#Concatting the df's
temp = pd.concat(temp, ignore_index=True, sort = False)

#Merging onto analysis
df_analysis = pd.merge(left = df_analysis, right = temp[['date', 'ID', 't']], left_on =['date', 'ID'], right_on = ['date', 'ID'])

#### Drop sector variables

criteria = 0.2

df_analysis.drop(['sector_management_staff', 'sector_trade_service', 'sector_sales_communication', 'sector_teaching', 'sector_social_health', 'sector_other'], axis = 1, inplace=True)

#### Transform relevant columns to abs change exept those with M_ and ID_, date and t

df_analysis['target_actual'] = df_analysis.groupby(['ID'])['target_actual'].diff()

### Lagged variables

df_analysis['target_lag'] = df_analysis.groupby(['ID'])['target_actual'].shift(1)

NB! As this is Quarterly data one year lag is 4 quarters

df_analysis['target_4_lag'] = df_analysis.groupby(['ID'])['target_actual'].shift(4)

#### Create new variables with 3 month (1 quarter) lag of jobrate

columns_1q_lag = ['jobs', 'sector_information_technology', 'sector_engineering_technology', 'sector_industry_craft','sector_office_finance']

for colname in columns_1q_lag:
    df_analysis[str(colname + '_1_lag')] = df_analysis.groupby(['ID'])[colname].shift(1)

#### Dropping some GT variables

# Dropping some GT's
drop_list = ['GT_SE_6', 'GT_SE_8', 'GT_SE_10', 'GT_SE_11', 'GT_DK_1']

df_analysis.drop(drop_list, axis = 1, inplace=True)

#### Create new variables with 1 month lag of GT

columns_1q_lag = ['GT_SE_0', 'GT_SE_1', 'GT_SE_2', 'GT_SE_5', 'GT_SE_7', 'GT_SE_9']
for colname in columns_1q_lag:
    df_analysis[str(colname + '_1_lag')] = df_analysis.groupby(['ID'])[colname].shift(1)

#### Month dummies for season effects

df_analysis['quarter'] = pd.DatetimeIndex(df_analysis['date']).month.astype(str)

#### Creating dummies from categorial variables - remember to drop the reference category (done after change is constructed)

df_analysis = pd.get_dummies(df_analysis, prefix=['ID','Q'], prefix_sep='_', columns=['ID', 'quarter']).copy()

#### Drop na

df_analysis.dropna(inplace=True)

df_analysis.date.max()

df_analysis.date.min()

#### Adding interaction terms

Polynominal features - To be deleted later

#df_analysis = add_poly_terms(df = df_analysis, 
#                            poly_columns = ['target_actual', 'GT_0', 'GT_1', 'GT_2', 'GT_3', 'GT_4', 'GT_5', 'GT_6', 'GT_7', 'GT_8', 'GT_9', 'GT_10', 'GT_11', 'GT_12', 'GT_13', 'GT_14', 'GT_15', 'GT_16', 'GT_17', 'GT_18', 'GT_19', 'target_lag', 'jobs', 'sector_information_technology', 'sector_engineering_technology', 'sector_management_staff', 'sector_trade_service', 'sector_industry_craft', 'sector_sales_communication', 'sector_teaching', 'sector_office_finance', 'sector_social_health', 'sector_other'])

#df_analysis.dropna(inplace=True)

Adding interaction terms by regions and all variables

# relevant interaction variables
interaction_1 = ['target_lag', 'target_4_lag'] 
# 'sector_information_technology', 'sector_engineering_technology', 'sector_management_staff', 'sector_trade_service', 'sector_industry_craft', 'sector_sales_communication', 'sector_teaching', 'sector_office_finance', 'sector_social_health', 'sector_other'

# get list of all ID area 
interaction_2 = [item for item in df_analysis if item.startswith('ID_')]

for var1 in interaction_1:
    for var2 in interaction_2:
        name = var1 + "*" + var2
        df_analysis[name] = pd.Series(df_analysis[var1] * df_analysis[var2], name=name)

#### Drop variables to not end up in dummytrap

df_analysis = df_analysis.drop(['ID_Blekinge', 'Q_1'], axis = 1)

df_analysis.drop(interaction_1, axis = 1, inplace=True)

df_analysis.sort_index(axis=1, inplace=True)

#### To Quarter data

df_analysis['date'] = df_analysis.set_index('date').index.to_period('Q')

df_analysis['date'].min()

df_analysis['date'].max()

#### Subset for period to match DK

df_analysis = df_analysis[df_analysis.date >= '01-01-2008']

df_analysis.columns

#### Setting window size

window = 11
testsize = 1
valsize = 1
rolling_window = True

## Lasso

### Test/train data split

Data must be split non-randomly as it needs to adhere to the underlying time structure. Two distinct approaches:

1. Rolling window (fixed length)
1. Expanding window (initial length that increases with each iteration)

Each model must be run as a loop over the test/train splits. Thus, we will have multiple test/train splits for both rolling window and expanding window.

X_train, X_val, X_test, y_train, y_val, y_test, y_dates = test_train_split_Q(df = df_analysis, window = window, testsize=testsize, valsize = valsize,
                                                                  y_col='target_actual', rolling_window = rolling_window, df_output= False, geo_count = 20)

with open('results/final/y_dates.pickle', 'wb') as handle:
    pickle.dump(y_dates, handle, protocol= pickle.HIGHEST_PROTOCOL)

### Pre-processing

- Standardizing

Standardizing features for each window

for win in X_train.keys():
    sc = StandardScaler()
    X_train[win] = sc.fit_transform(X_train[win])
    X_val[win] = sc.transform(X_val[win])
    X_test[win] = sc.transform(X_test[win])

### Training the models

#### Hyperparameter space - random

alphas = np.logspace(-8, 8, num = 10000) #random.sample(list(np.logspace(-10,5, num = 15000)), k = 15000)
n_components = list(np.arange(0.6, 0.95, 0.05).round(2))#[0.60, 0.70, 0.80, 0.90]
params = [(alpha) for alpha in alphas]
print('Number of param sets: '+ str(len(params)))

#### Inner loop - training hyperparameter on validation

##### On random space

results_lasso_mp = tuning_window_mp(X_fit = X_train, y_fit = y_train, X_test = X_val, y_test = y_val, params = params, n_components = n_components, model_str = 'lasso')

with open('results/final/lasso/results_mp.pickle', 'wb') as handle:
    pickle.dump(results_lasso_mp, handle, protocol= pickle.HIGHEST_PROTOCOL)

#### Outer loop - fitting on train / test split

Importing stored results:

with open('results/final/lasso/results_mp.pickle', 'rb') as handle:
    results_lasso_opt = pickle.load(handle)

##### On full sample

Reloading data and concatting:

X_train, X_val, X_test, y_train, y_val, y_test, dates = test_train_split_Q(df = df_analysis, window = window, testsize=testsize, valsize = valsize,
                                                                  y_col='target_actual', rolling_window = rolling_window, df_output= False, geo_count = 20)

#Concatting val and train
for win in X_train.keys():
    X_train[win] = np.concatenate((X_train[win], X_val[win])).copy()
    y_train[win] = np.concatenate((y_train[win], y_val[win])).copy()

Standardizing features for each window

for win in X_train.keys():
    sc = StandardScaler()
    X_train[win] = sc.fit_transform(X_train[win])
    X_test[win] = sc.transform(X_test[win])

results_final = final_model(inner_results=results_lasso_opt, X_fit = X_train, y_fit = y_train, X_test = X_test, y_test = y_test, model_str = 'lasso')

#### Exporting final results

with open('results/final/lasso/results_final.pickle', 'wb') as handle:
    pickle.dump(results_final, handle, protocol= pickle.HIGHEST_PROTOCOL)

temp = []
for key in results_final.keys():
    temp.append(results_final[key]['best_rmse'][1])
np.mean(temp)

## Ridge

X_train, X_val, X_test, y_train, y_val, y_test, y_dates = test_train_split_Q(df = df_analysis, window = window, testsize=testsize, valsize = valsize,
                                                                  y_col='target_actual', rolling_window = rolling_window, df_output= False, geo_count = 20)

### Pre-processing

- Standardizing

Standardizing features for each window

for win in X_train.keys():
    sc = StandardScaler()
    X_train[win] = sc.fit_transform(X_train[win])
    X_val[win] = sc.transform(X_val[win])
    X_test[win] = sc.transform(X_test[win])

### Training the models

#### Hyperparameter space - random

alphas = np.logspace(-8, 8, num = 10000) #random.sample(list(np.logspace(-10,5, num = 15000)), k = 15000)
n_components = list(np.arange(0.6, 0.95, 0.05).round(2))#[0.60, 0.70, 0.80, 0.90]
params = [(alpha) for alpha in alphas]
print('Number of param sets: '+ str(len(params)))

#### Inner loop - training hyperparameter on validation

##### On random space

results_mp = tuning_window_mp(X_fit = X_train, y_fit = y_train, X_test = X_val, y_test = y_val, params = params, n_components = n_components, model_str = 'ridge')

with open('results/final/ridge/results_mp.pickle', 'wb') as handle:
    pickle.dump(results_mp, handle, protocol= pickle.HIGHEST_PROTOCOL)

#### Outer loop - fitting on train / test split

Importing stored results:

with open('results/final/ridge/results_mp.pickle', 'rb') as handle:
    results_opt = pickle.load(handle)

##### On full sample

Reloading data and concatting:

X_train, X_val, X_test, y_train, y_val, y_test, y_dates = test_train_split_Q(df = df_analysis, window = window, testsize=testsize, valsize = valsize,
                                                                  y_col='target_actual', rolling_window = rolling_window, df_output= False, geo_count = 20)

#Concatting val and train
for win in X_train.keys():
    X_train[win] = np.concatenate((X_train[win], X_val[win])).copy()
    y_train[win] = np.concatenate((y_train[win], y_val[win])).copy()

Standardizing features for each window

for win in X_train.keys():
    sc = StandardScaler()
    X_train[win] = sc.fit_transform(X_train[win])
    X_test[win] = sc.transform(X_test[win])

results_final = final_model(inner_results=results_opt, X_fit = X_train, y_fit = y_train, X_test = X_test, y_test = y_test, model_str = 'ridge')

#### Exporting final results

with open('results/final/ridge/results_final.pickle', 'wb') as handle:
    pickle.dump(results_final, handle, protocol= pickle.HIGHEST_PROTOCOL)

temp = []
for key in results_final.keys():
    temp.append(results_final[key]['best_rmse'][1])
np.mean(temp)

## Elastic net

X_train, X_val, X_test, y_train, y_val, y_test, y_dates = test_train_split_Q(df = df_analysis, window = window, testsize=testsize, valsize = valsize,
                                                                  y_col='target_actual', rolling_window = rolling_window, df_output= False, geo_count = 20)

### Pre-processing

- Standardizing

Standardizing features for each window

for win in X_train.keys():
    sc = StandardScaler()
    X_train[win] = sc.fit_transform(X_train[win])
    X_val[win] = sc.transform(X_val[win])
    X_test[win] = sc.transform(X_test[win])

### Training the models

#### Hyperparameter space - random

alphas = np.logspace(-8,8, num = 10000) #2000
n_components= [0.60, 0.65, 0.70, 0.75, 0.80, 0.85, 0.90]
l1_ratio = list(np.arange(0.01,0.99,0.05)) ## 0.0 cannot be included due to a bug in the code
params = [(alpha, ratio) for alpha in alphas for ratio in l1_ratio]
print('Number of param sets: '+ str(len(params)))


#### Inner loop - training hyperparameter on validation

##### On random space

results_mp = tuning_window_mp(X_fit = X_train, y_fit = y_train, X_test = X_val, y_test = y_val, params = params, n_components = n_components, model_str = 'elasticnet')

with open('results/final/elastic/results_mp.pickle', 'wb') as handle:
    pickle.dump(results_mp, handle, protocol= pickle.HIGHEST_PROTOCOL)

#### Outer loop - fitting on train / test split

Importing stored results:

with open('results/final/elastic/results_mp.pickle', 'rb') as handle:
    results_opt = pickle.load(handle)

##### On full sample

Reloading data and concatting:

X_train, X_val, X_test, y_train, y_val, y_test, y_dates = test_train_split_Q(df = df_analysis, window = window, testsize=testsize, valsize = valsize,
                                                                  y_col='target_actual', rolling_window = rolling_window, df_output= False, geo_count = 20)

#Concatting val and train
for win in X_train.keys():
    X_train[win] = np.concatenate((X_train[win], X_val[win])).copy()
    y_train[win] = np.concatenate((y_train[win], y_val[win])).copy()

Standardizing features for each window

for win in X_train.keys():
    sc = StandardScaler()
    X_train[win] = sc.fit_transform(X_train[win])
    X_test[win] = sc.transform(X_test[win])

results_final = final_model(inner_results=results_opt, X_fit = X_train, y_fit = y_train, X_test = X_test, y_test = y_test, model_str = 'elasticnet')

#### Exporting final results

with open('results/final/elastic/results_final.pickle', 'wb') as handle:
    pickle.dump(results_final, handle, protocol= pickle.HIGHEST_PROTOCOL)

temp = []
for key in results_final.keys():
    temp.append(results_final[key]['best_rmse'][1])
np.mean(temp)

## ML - Data and preprocessing - tree-based

### Import data frame with adjusted here

### Subset data 

df_analysis = pd.read_csv('data/descriptive/df_descriptive.csv', parse_dates=['date'])

df_analysis = df_analysis[df_analysis.country == 'SE']

df_analysis = df_analysis[~(df_analysis.ID == 'Jämtland')]

df_analysis = df_analysis.drop(['GT_NO_0', 'GT_NO_1', 'GT_NO_2', 'GT_NO_6', 'GT_NO_7', 'GT_NO_8', 'GT_NO_10', 'GT_NO_11',  
                                'GT_DK_0', 'GT_DK_2', 'GT_DK_3', 'GT_DK_4', 'GT_DK_5', 'GT_DK_6', 'GT_DK_7', 'GT_DK_8',
                                'GT_DK_9', 'GT_DK_10', 'GT_DK_11', 'GT_DK_12', 'GT_DK_13', 'GT_DK_14',
                                'GT_DK_15', 'GT_DK_16', 'GT_DK_17', 'GT_DK_18', 'GT_DK_19', 'GT_DK_20', 'GT_DK_21',
                                'country'], axis = 1)

print(list(df_analysis))

df_analysis.date.min()

df_analysis.date.max()

### Initial preprocessing and feature construction

- Create dummies 
- Create interaction terms

Overall monthly time trend variable, $t=1,2...,T$ within `ID` variable:

#Temp container
temp = {}

for i in df_analysis['ID'].unique():
    temp[i] = df_analysis[df_analysis['ID']==i]
    temp[i]['t'] = range(1, len(temp[i]['ID'])+1)

#Concatting the df's
temp = pd.concat(temp, ignore_index=True, sort = False)

#Merging onto analysis
df_analysis = pd.merge(left = df_analysis, right = temp[['date', 'ID', 't']], left_on =['date', 'ID'], right_on = ['date', 'ID'])

#### Drop sector variables

criteria = 0.2

df_analysis.drop(['sector_management_staff', 'sector_trade_service', 'sector_sales_communication', 'sector_teaching', 'sector_social_health', 'sector_other'], axis = 1, inplace=True)

#### Transform relevant columns to abs change exept those with M_ and ID_, date and t

df_analysis['target_actual'] = df_analysis.groupby(['ID'])['target_actual'].diff()

### Lagged variables

df_analysis['target_lag'] = df_analysis.groupby(['ID'])['target_actual'].shift(1)

NB! As this is Quarterly data one year lag is 4 quarters

df_analysis['target_4_lag'] = df_analysis.groupby(['ID'])['target_actual'].shift(4)

#### Create new variables with 3 month (1 quarter) lag of jobrate

columns_1q_lag = ['jobs', 'sector_information_technology', 'sector_engineering_technology', 'sector_industry_craft','sector_office_finance']

for colname in columns_1q_lag:
    df_analysis[str(colname + '_1_lag')] = df_analysis.groupby(['ID'])[colname].shift(1)

#### Dropping some GT variables

# Dropping some GT's
drop_list = ['GT_SE_6', 'GT_SE_8', 'GT_SE_10', 'GT_SE_11', 'GT_DK_1']

df_analysis.drop(drop_list, axis = 1, inplace=True)

#### Create new variables with 1 month lag of GT

columns_1q_lag = ['GT_SE_0', 'GT_SE_1', 'GT_SE_2', 'GT_SE_5', 'GT_SE_7', 'GT_SE_9']
for colname in columns_1q_lag:
    df_analysis[str(colname + '_1_lag')] = df_analysis.groupby(['ID'])[colname].shift(1)

#### Month dummies for season effects

df_analysis['quarter'] = pd.DatetimeIndex(df_analysis['date']).month.astype(str)

#### Creating dummies from categorial variables - remember to drop the reference category (done after change is constructed)

df_analysis = pd.get_dummies(df_analysis, prefix=['ID','Q'], prefix_sep='_', columns=['ID', 'quarter']).copy()

#### Drop na

df_analysis.dropna(inplace=True)

df_analysis.date.max()

df_analysis.date.min()

#### Adding interaction terms

Polynominal features - To be deleted later

#df_analysis = add_poly_terms(df = df_analysis, 
#                            poly_columns = ['target_actual', 'GT_0', 'GT_1', 'GT_2', 'GT_3', 'GT_4', 'GT_5', 'GT_6', 'GT_7', 'GT_8', 'GT_9', 'GT_10', 'GT_11', 'GT_12', 'GT_13', 'GT_14', 'GT_15', 'GT_16', 'GT_17', 'GT_18', 'GT_19', 'target_lag', 'jobs', 'sector_information_technology', 'sector_engineering_technology', 'sector_management_staff', 'sector_trade_service', 'sector_industry_craft', 'sector_sales_communication', 'sector_teaching', 'sector_office_finance', 'sector_social_health', 'sector_other'])

#df_analysis.dropna(inplace=True)

Adding interaction terms by regions and all variables

# # relevant interaction variables
# interaction_1 = ['target_lag', 'target_4_lag'] 
# # 'sector_information_technology', 'sector_engineering_technology', 'sector_management_staff', 'sector_trade_service', 'sector_industry_craft', 'sector_sales_communication', 'sector_teaching', 'sector_office_finance', 'sector_social_health', 'sector_other'

# # get list of all ID area 
# interaction_2 = [item for item in df_analysis if item.startswith('ID_')]

# for var1 in interaction_1:
#     for var2 in interaction_2:
#         name = var1 + "*" + var2
#         df_analysis[name] = pd.Series(df_analysis[var1] * df_analysis[var2], name=name)

#### Drop variables to not end up in dummytrap

df_analysis = df_analysis.drop(['ID_Blekinge', 'Q_1'], axis = 1)

# df_analysis.drop(interaction_1, axis = 1, inplace=True)

df_analysis.sort_index(axis=1, inplace=True)

#### To Quarter data

df_analysis['date'] = df_analysis.set_index('date').index.to_period('Q')

df_analysis['date'].min()

df_analysis['date'].max()

#### Subset for period to match DK

df_analysis = df_analysis[df_analysis.date >= '01-01-2008']

#### Setting window size

window = 11
testsize = 1
valsize = 1
rolling_window = True

## Random forest

X_train, X_val, X_test, y_train, y_val, y_test, y_dates = test_train_split_Q(df = df_analysis, window = window, testsize=testsize, valsize = valsize,
                                                                  y_col='target_actual', rolling_window = rolling_window, df_output= False, geo_count = 20)

### Pre-processing

- Standardizing

Standardizing features for each window

for win in X_train.keys():
    sc = StandardScaler()
    X_train[win] = sc.fit_transform(X_train[win])
    X_val[win] = sc.transform(X_val[win])
    X_test[win] = sc.transform(X_test[win])

### Training the models

#### Hyperparameter space - random

# Website https://towardsdatascience.com/hyperparameter-tuning-the-random-forest-in-python-using-scikit-learn-28d2aa77dd74

# parameter specification
n_components = [0.9]

#range(1, X_train[1].shape[1] +1)
# Number of trees in random forest
n_estimators = [*range(50, 500, 10)]
# Number of features to consider at every split
max_features = ['auto', "sqrt"] # Consider whether this should be set to 'auto as PCA should do its job'
# Maximum number of levels in tree
max_depth = [*range(3, 11, 1), *range(20, 100, 20)]
max_depth.append(None)
# Minimum number of samples required to split a node
min_samples_split = [2]
# Minimum number of samples required at each leaf node
min_samples_leaf = [1, 3, 5]
# Method of selecting samples for training each tree
# https://gdcoder.com/random-forest-regression-model-explained-in-depth-part-2-python-code-snippet-using-sklearn/

# Create the random grid
d = {
    'n_estimators': n_estimators,
    'max_depth': max_depth,
    'max_features': max_features,
    'min_samples_split': min_samples_split,
     'min_samples_leaf':min_samples_leaf
    }

params = list(d.values())

params = list(itertools.product(*params))

print(len(params))


#### Inner loop - training hyperparameter on validation

##### On random space

results_mp = tuning_window(X_fit = X_train, y_fit = y_train, X_test = X_val, y_test = y_val, params = params, n_components = n_components, model_str = 'randomforest')

with open('results/final/randomforest/results_noint.pickle', 'wb') as handle:
    pickle.dump(results_mp, handle, protocol= pickle.HIGHEST_PROTOCOL)

#### Outer loop - fitting on train / test split

Importing stored results:

with open('results/final/randomforest/results_noint.pickle', 'rb') as handle:
    results_opt = pickle.load(handle)

##### On full sample

Reloading data and concatting:

X_train, X_val, X_test, y_train, y_val, y_test, y_dates = test_train_split_Q(df = df_analysis, window = window, testsize=testsize, valsize = valsize,
                                                                  y_col='target_actual', rolling_window = rolling_window, df_output= False, geo_count = 20)

#Concatting val and train
for win in X_train.keys():
    X_train[win] = np.concatenate((X_train[win], X_val[win])).copy()
    y_train[win] = np.concatenate((y_train[win], y_val[win])).copy()

Standardizing features for each window

for win in X_train.keys():
    sc = StandardScaler()
    X_train[win] = sc.fit_transform(X_train[win])
    X_test[win] = sc.transform(X_test[win])

results_final = final_model(inner_results=results_opt, X_fit = X_train, y_fit = y_train, X_test = X_test, y_test = y_test, model_str = 'randomforest')

#### Exporting final results

with open('results/final/randomforest/results_final_noint.pickle', 'wb') as handle:
    pickle.dump(results_final, handle, protocol= pickle.HIGHEST_PROTOCOL)

temp = []
for key in results_final.keys():
    temp.append(results_final[key]['best_rmse'][1])
np.mean(temp)

## XGboost

X_train, X_val, X_test, y_train, y_val, y_test, y_dates = test_train_split_Q(df = df_analysis, window = window, testsize=testsize, valsize = valsize,
                                                                  y_col='target_actual', rolling_window = rolling_window, df_output= False, geo_count = 20)

### Pre-processing

- Standardizing

Standardizing features for each window

for win in X_train.keys():
    sc = StandardScaler()
    X_train[win] = sc.fit_transform(X_train[win])
    X_val[win] = sc.transform(X_val[win])
    X_test[win] = sc.transform(X_test[win])

### Training the models

#### Hyperparameter space - random

# Website https://towardsdatascience.com/hyperparameter-tuning-the-random-forest-in-python-using-scikit-learn-28d2aa77dd74

# parameter specification
n_components = [0.9]


colsample_bytree = [0.3, 0.5, 0.7, 0.9, 1]#is the subsample ratio of columns when constructing each tree. Subsampling will occur once in every boosting iteration. This number ranges from 0 to 1.
#learning_rate is the step size shrinkage and is used to prevent overfitting. This number ranges from 0 to 1.
# Maximum number of levels in tree
# First try: colsample_bytree = [0.3, 0.4, 0.5, 0.6, 0.7, 0.8, 0.9, 1]

max_depth = [*range(3, 11, 1), *range(20, 100, 20)]

#max_depth = [*range(10, 300, 10)]
# first try max_depth = [*range(50, 500, 20)]

n_estimators = [*range(50, 500, 10)]#is the number of boosted trees to fit
# first try n_estimators = [*range(50, 500, 20)]

gamma = [0]

subsample = [0.5, 0.75, 1]

min_child_weight = [1, 3, 5]

# Create the random grid
d = {
    'n_estimators': n_estimators,
    'max_depth': max_depth,
    'colsample_bytree': colsample_bytree,
    'gamma' : gamma,
    'subsample' : subsample,
    'min_child_weight' : min_child_weight
    }

params = list(d.values())

params = list(itertools.product(*params))

print(len(params))

random.seed(1)
params = random.sample(params, 10000)

#### Inner loop - training hyperparameter on validation

##### On random space

results_mp = tuning_window_mp(X_fit = X_train, y_fit = y_train, X_test = X_val, y_test = y_val, params = params, n_components = n_components, model_str = 'xgboost')

with open('results/final/xgboost/results_noint.pickle', 'wb') as handle:
    pickle.dump(results_mp, handle, protocol= pickle.HIGHEST_PROTOCOL)

#### Outer loop - fitting on train / test split

##### On full sample

Reloading data and concatting:

with open('results/final/xgboost/results_noint.pickle', 'rb') as handle:
    results_opt = pickle.load(handle)
    

X_train, X_val, X_test, y_train, y_val, y_test, y_dates = test_train_split_Q(df = df_analysis, window = window, testsize=testsize, valsize = valsize,
                                                                  y_col='target_actual', rolling_window = rolling_window, df_output= False, geo_count = 20)

#Concatting val and train
for win in X_train.keys():
    X_train[win] = np.concatenate((X_train[win], X_val[win])).copy()
    y_train[win] = np.concatenate((y_train[win], y_val[win])).copy()

Standardizing features for each window

for win in X_train.keys():
    sc = StandardScaler()
    X_train[win] = sc.fit_transform(X_train[win])
    X_test[win] = sc.transform(X_test[win])

results_final = final_model(inner_results=results_opt, X_fit = X_train, y_fit = y_train, X_test = X_test, y_test = y_test, model_str = 'xgboost')

#### Exporting final results

with open('results/final/xgboost/results_final_noint.pickle', 'wb') as handle:
    pickle.dump(results_final, handle, protocol= pickle.HIGHEST_PROTOCOL)

temp = []
for key in results_final.keys():
    temp.append(results_final[key]['best_rmse'][1])
np.mean(temp)
